REFACTOR/#5: 전체적인 코드 리팩토링
