+++ app/src/main/java/com/sopt/now/compose/ui/screens/search/SearchScreen.kt
@@ -1,29 +1,16 @@
 package com.sopt.now.compose.ui.screens.search
 
 import androidx.activity.compose.BackHandler
-import androidx.activity.compose.LocalOnBackPressedDispatcherOwner
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.icons.filled.Search
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.lifecycle.LifecycleOwner
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
-import com.sopt.now.compose.R
-import com.sopt.now.compose.SoptApplication.Companion.NAVIGATE_BACK_PRESSED_KEY
+import com.sopt.now.compose.MainActivity.Companion.NAVIGATE_BACK_PRESSED_KEY
 import com.sopt.now.compose.ext.putDataAtPreviousBackStackEntry
-import com.sopt.now.compose.ext.removePreviousBackStacks
 import com.sopt.now.compose.ui.SoptBottomNavigation
-import com.sopt.now.compose.ui.navigation.LoginDestination
-import com.sopt.now.compose.ui.navigation.NavigationDestination
 
 
 /**
