+++ app/src/main/java/com/sopt/now/compose/ui/screens/signup/SignUpScreen.kt
@@ -19,9 +19,9 @@ import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
+import com.sopt.now.compose.MainActivity.Companion.NAVIGATE_SIGNUP_KEY
 import com.sopt.now.compose.MainActivity.Companion.printToastMessage
 import com.sopt.now.compose.R
-import com.sopt.now.compose.SoptApplication.Companion.NAVIGATE_SIGNUP_KEY
 import com.sopt.now.compose.ui.composables.ButtonComposable
 import com.sopt.now.compose.ui.composables.TextFieldWithTitleComposable
 import com.sopt.now.compose.ext.putDataAtPreviousBackStackEntry
